+++ R/install.R
@@ -37,9 +37,11 @@
 #' @param force Whether to install Hugo even if it has already been installed.
 #'   This may be useful when upgrading Hugo (if you use Homebrew, run the
 #'   command \command{brew update && brew upgrade} instead).
+#' @param extended Whether to use extended version of Hugo that has SCSS/SASS support.
+#'   You only need the extended version if you want to edit SCSS/SASS.
 #' @export
 install_hugo = function(
-  version = 'latest', use_brew = Sys.which('brew') != '', force = FALSE
+  version = 'latest', use_brew = Sys.which('brew') != '', force = FALSE, extended = TRUE
 ) {
 
   if (Sys.which('hugo') != '' && !force) {
@@ -74,6 +76,13 @@ install_hugo = function(
   version = gsub('^[vV]', '', version)  # pure version number
   version2 = as.numeric_version(version)
   bit = if (grepl('64', Sys.info()[['machine']])) '64bit' else '32bit'
+  if (extended) {
+    if (bit != '64bit') stop('The extended version of Hugo is only available on 64-bit platforms')
+    if (version2 < '0.43') {
+      if (!missing(extended)) stop('Only Hugo >= v0.43 provides the extended version')
+      extended = FALSE
+    }
+  }
   base = sprintf('https://github.com/gohugoio/hugo/releases/download/v%s/', version)
   owd = setwd(tempdir())
   on.exit(setwd(owd), add = TRUE)
@@ -81,7 +90,9 @@ install_hugo = function(
 
   download_zip = function(OS, type = 'zip') {
     if (is.null(local_file)) {
-      zipfile = sprintf('hugo_%s_%s-%s.%s', version, OS, bit, type)
+      zipfile = sprintf(
+        'hugo_%s%s_%s-%s.%s', ifelse(extended, 'extended_', ''), version, OS, bit, type
+      )
       xfun::download_file(paste0(base, zipfile), zipfile, mode = 'wb')
     } else {
       zipfile = local_file